Separate libevdev, pynput solutions into modules
